explore alternative approaches

+++ app/src/main/java/com/badlogic/androidgames/framework/Graphics.java
@@ -10,20 +10,20 @@ public interface Graphics {
 
     public void clear(int color);
 
-    public void drawPixel(int x, int y, int color);
+    public void drawPixel(float x, float y, int color);
 
-    public void drawLine(int x, int y, int x2, int y2, int color);
+    public void drawLine(float x, float y, float x2, float y2, int color);
 
-    public void drawRect(int x, int y, int width, int height, int color);
+    public void drawRect(float x, float y, float width, float height, int color);
 
     public void drawPixmap(Pixmap pixmap, int x, int y, int srcX, int srcY,
             int srcWidth, int srcHeight);
 
-    public void drawPixmap(Pixmap pixmap, int x, int y);
+    public void drawPixmap(Pixmap pixmap, float x, float y);
 
     public int getWidth();
 
     public int getHeight();
 
-    public void drawCircle(int x, int y, float radius, int color, int alpha);
+    public void drawCircle(float x, float y, float radius, int color, int alpha);
 }
