+++ app/src/main/java/com/badlogic/androidgames/framework/impl/AndroidGraphics.java
@@ -13,6 +13,7 @@ import android.graphics.Color;
 import android.graphics.Paint;
 import android.graphics.Paint.Style;
 import android.graphics.Rect;
+import android.graphics.RectF;
 
 import com.badlogic.androidgames.framework.Graphics;
 import com.badlogic.androidgames.framework.Pixmap;
@@ -82,13 +83,13 @@ public class AndroidGraphics implements Graphics {
     }
 
     @Override
-    public void drawPixel(int x, int y, int color) {
+    public void drawPixel(float x, float y, int color) {
         paint.setColor(color);
         canvas.drawPoint(x, y, paint);
     }
 
     @Override
-    public void drawLine(int x, int y, int x2, int y2, int color) {
+    public void drawLine(float x, float y, float x2, float y2, int color) {
         paint.setColor(color);
         paint.setAlpha(255);
         paint.setStyle(Style.STROKE);
@@ -97,7 +98,7 @@ public class AndroidGraphics implements Graphics {
     }
 
     @Override
-    public void drawRect(int x, int y, int width, int height, int color) {
+    public void drawRect(float x, float y, float width, float height, int color) {
         paint.setColor(color);
         paint.setStyle(Style.FILL);
         canvas.drawRect(x, y, x + width - 1, y + width - 1, paint);
@@ -116,12 +117,11 @@ public class AndroidGraphics implements Graphics {
         dstRect.right = x + srcWidth - 1;
         dstRect.bottom = y + srcHeight - 1;
 
-        canvas.drawBitmap(((AndroidPixmap) pixmap).bitmap, srcRect, dstRect,
-                null);
+        canvas.drawBitmap(((AndroidPixmap) pixmap).bitmap, srcRect, dstRect,null);
     }
     
     @Override
-    public void drawPixmap(Pixmap pixmap, int x, int y) {
+    public void drawPixmap(Pixmap pixmap, float x, float y) {
         canvas.drawBitmap(((AndroidPixmap)pixmap).bitmap, x, y, null);
     }
 
@@ -136,7 +136,7 @@ public class AndroidGraphics implements Graphics {
     }
 
     @Override
-    public void drawCircle(int x, int y, float radius, int color, int alpha){
+    public void drawCircle(float x, float y, float radius, int color, int alpha){
         paint.setColor(color);
         paint.setAlpha(alpha/2);
         paint.setStyle(Style.FILL);
